refactor(profile): profile_tab_page.dart 분해 (2628→1755줄)
대규모 ProfileTabPageState를 그대로 두고 도메인별 메소드를
extension으로 part 파일에 분리. 코드 변경 0건 (메소드 이동만).
extension은 같은 라이브러리이므로 private 멤버 모두 접근 가능.

신설:
- lib/widgets/profile/profile_person_overview.dart (~400줄, part)
  · _openProfilePersonOverview - 인물별 학습 진행도 다이얼로그
- lib/widgets/profile/profile_intercessory_prayer.dart (~225줄, part)
  · _buildIntercessoryPrayerErrorCard / _buildIntercessoryPrayerEmptyCard
  · _buildIntercessoryPrayerItemCard / _intercessoryPrayerFab
- lib/widgets/profile/profile_helpers.dart (~260줄, part)
  · _profileNetworkAvatar, _profileTinyIconButton
  · _profileTestamentToggleButton, _profileTopStatCard
  · _profilePersonProgressRow

부모 파일:
- 해당 메소드 블록 875줄 제거
- part 'profile/...' 디렉티브 3개 추가
- 2628 → 1755줄 (-33%)

flutter analyze: 0 issues
flutter test: 107 tests pass
