+++ lib/widgets/profile_tab_page.dart
@@ -30,6 +30,12 @@ import 'share_id_input_dialog.dart';
 import 'story_home_styles.dart';
 import 'sub_page_scaffold.dart';
 
+// 화면 코드를 도메인별로 part 파일로 분리.
+// 각 part 파일은 ProfileTabPageState에 대한 extension으로 메소드를 추가한다.
+part 'profile/profile_helpers.dart';
+part 'profile/profile_intercessory_prayer.dart';
+part 'profile/profile_person_overview.dart';
+
 /// 프로필 탭 페이지 (프로필 정보 + 인물 진행도 + 노트/말씀/중보기도 미리보기).
 ///
 /// 외부 콜백:
@@ -1742,888 +1748,6 @@ class ProfileTabPageState extends ConsumerState<ProfileTabPage> {
           ),
         ),
       ),
-    );
-  }
-
-  Widget _buildIntercessoryPrayerErrorCard() {
-    return Container(
-      decoration: BoxDecoration(
-        color: const Color(0x18A63F2D),
-        borderRadius: BorderRadius.circular(14),
-        border: Border.all(color: const Color(0x66A63F2D), width: 1),
-      ),
-      padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 14),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Text(
-            _intercessoryPrayerError!,
-            textAlign: TextAlign.center,
-            style: const TextStyle(
-              color: Color(0xFF7E3426),
-              fontSize: 13.0,
-              fontWeight: FontWeight.w800,
-              height: 1.35,
-            ),
-          ),
-          const SizedBox(height: 10),
-          _profileMiniActionButton(
-            label: '다시 불러오기',
-            onTap: _loadIntercessoryPrayerPage,
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildIntercessoryPrayerEmptyCard({required bool enabled}) {
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        final isVeryCompact = constraints.maxHeight < 150;
-        final isCompact = constraints.maxHeight < 180;
-        final buttonSize = isVeryCompact ? 32.0 : (isCompact ? 38.0 : 44.0);
-        final iconSize = isVeryCompact ? 20.0 : (isCompact ? 24.0 : 26.0);
-        final spacing = isVeryCompact ? 4.0 : (isCompact ? 6.0 : 8.0);
-        final fontSize = isVeryCompact ? 10.4 : (isCompact ? 11.2 : 12.3);
-        return Material(
-          color: Colors.transparent,
-          child: InkWell(
-            onTap: enabled ? _promptAddIntercessoryPrayer : null,
-            borderRadius: BorderRadius.circular(16),
-            child: Center(
-              child: SingleChildScrollView(
-                physics: const NeverScrollableScrollPhysics(),
-                child: Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 8),
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      Container(
-                        width: buttonSize,
-                        height: buttonSize,
-                        decoration: BoxDecoration(
-                          shape: BoxShape.circle,
-                          gradient: enabled
-                              ? const LinearGradient(
-                                  begin: Alignment.topLeft,
-                                  end: Alignment.bottomRight,
-                                  colors: [
-                                    Color(0xFFD99F4A),
-                                    Color(0xFFB26B28),
-                                  ],
-                                )
-                              : const LinearGradient(
-                                  begin: Alignment.topLeft,
-                                  end: Alignment.bottomRight,
-                                  colors: [
-                                    Color(0xFFD7CCB9),
-                                    Color(0xFFB6A38A),
-                                  ],
-                                ),
-                          boxShadow: const [
-                            BoxShadow(
-                              color: Color(0x22000000),
-                              blurRadius: 7,
-                              offset: Offset(0, 2),
-                            ),
-                          ],
-                        ),
-                        child: Icon(
-                          Icons.add_rounded,
-                          color: const Color(0xFFFDF8EE),
-                          size: iconSize,
-                        ),
-                      ),
-                      SizedBox(height: spacing),
-                      Text(
-                        enabled
-                            ? '다른 사람의 기도제목을 공유 받아\n함께 기도해요'
-                            : '로그인하면 다른 사람의 기도제목을\n함께 볼 수 있어요',
-                        textAlign: TextAlign.center,
-                        maxLines: 2,
-                        overflow: TextOverflow.ellipsis,
-                        style: TextStyle(
-                          color: const Color(0xFF5A4326),
-                          fontSize: fontSize,
-                          fontWeight: FontWeight.w800,
-                          height: 1.24,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  Widget _buildIntercessoryPrayerItemCard(IntercessoryPrayerItem item) {
-    final prayerText = (item.prayerRequest ?? '').trim().isEmpty
-        ? '아직 등록된 기도제목이 없어요.'
-        : item.prayerRequest!.trim();
-    return Container(
-      decoration: BoxDecoration(
-        color: const Color(0xC9F1E3CB),
-        borderRadius: BorderRadius.circular(14),
-        border: Border.all(color: const Color(0xAA8E6F48), width: 1.0),
-      ),
-      padding: const EdgeInsets.fromLTRB(10, 10, 8, 10),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          _profileNetworkAvatar(
-            nickname: item.nickname,
-            photoUrl: item.photoUrl,
-            size: 42,
-          ),
-          const SizedBox(width: 10),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Row(
-                  children: [
-                    Expanded(
-                      child: Text(
-                        item.nickname,
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                        style: const TextStyle(
-                          color: Color(0xFF452F1A),
-                          fontWeight: FontWeight.w900,
-                          fontSize: 13.4,
-                        ),
-                      ),
-                    ),
-                    const SizedBox(width: 6),
-                    Text(
-                      item.shareId,
-                      maxLines: 1,
-                      style: const TextStyle(
-                        color: Color(0xFF8A6A46),
-                        fontSize: 10.4,
-                        fontWeight: FontWeight.w800,
-                      ),
-                    ),
-                  ],
-                ),
-                const SizedBox(height: 5),
-                Text(
-                  prayerText,
-                  maxLines: 4,
-                  overflow: TextOverflow.ellipsis,
-                  style: const TextStyle(
-                    color: Color(0xFF5A4326),
-                    fontWeight: FontWeight.w700,
-                    fontSize: 12.6,
-                    height: 1.34,
-                  ),
-                ),
-              ],
-            ),
-          ),
-          const SizedBox(width: 6),
-          _profileTinyIconButton(
-            tooltip: '삭제',
-            onTap: () => _confirmDeleteIntercessoryPrayer(item),
-            icon: Icons.delete_outline_rounded,
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _intercessoryPrayerFab({required bool enabled}) {
-    return Material(
-      color: Colors.transparent,
-      elevation: 10,
-      shadowColor: const Color(0x33000000),
-      shape: const CircleBorder(),
-      child: InkWell(
-        onTap: enabled ? _promptAddIntercessoryPrayer : null,
-        customBorder: const CircleBorder(),
-        child: Ink(
-          width: 38,
-          height: 38,
-          decoration: BoxDecoration(
-            shape: BoxShape.circle,
-            gradient: enabled
-                ? const LinearGradient(
-                    begin: Alignment.topLeft,
-                    end: Alignment.bottomRight,
-                    colors: [Color(0xFFD99F4A), Color(0xFFB26B28)],
-                  )
-                : const LinearGradient(
-                    begin: Alignment.topLeft,
-                    end: Alignment.bottomRight,
-                    colors: [Color(0xFFD7CCB9), Color(0xFFB6A38A)],
-                  ),
-            border: Border.all(color: const Color(0xFFF2D8A6), width: 1.1),
-          ),
-          child: const Icon(
-            Icons.add_rounded,
-            color: Color(0xFFFDF8EE),
-            size: 21,
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _profileNetworkAvatar({
-    required String nickname,
-    required String? photoUrl,
-    double size = 42,
-  }) {
-    final initials = nickname.trim().isEmpty ? '?' : nickname.trim()[0];
-    final hasPhoto = (photoUrl ?? '').trim().isNotEmpty;
-    return Container(
-      width: size,
-      height: size,
-      decoration: BoxDecoration(
-        shape: BoxShape.circle,
-        gradient: const LinearGradient(
-          begin: Alignment.topLeft,
-          end: Alignment.bottomRight,
-          colors: [Color(0xFFEFD79B), Color(0xFFC88A3D)],
-        ),
-        border: Border.all(color: const Color(0xFF8C6743), width: 1.2),
-      ),
-      child: ClipOval(
-        child: hasPhoto
-            ? Image.network(
-                photoUrl!.trim(),
-                fit: BoxFit.cover,
-                errorBuilder: (_, __, ___) {
-                  return Center(
-                    child: Text(
-                      initials,
-                      style: TextStyle(
-                        color: const Color(0xFF4A331D),
-                        fontSize: size * 0.34,
-                        fontWeight: FontWeight.w900,
-                      ),
-                    ),
-                  );
-                },
-              )
-            : Center(
-                child: Text(
-                  initials,
-                  style: TextStyle(
-                    color: const Color(0xFF4A331D),
-                    fontSize: size * 0.34,
-                    fontWeight: FontWeight.w900,
-                  ),
-                ),
-              ),
-      ),
-    );
-  }
-
-  Widget _profileTinyIconButton({
-    required String tooltip,
-    required VoidCallback onTap,
-    required IconData icon,
-  }) {
-    return Tooltip(
-      message: tooltip,
-      child: Material(
-        color: Colors.transparent,
-        child: InkWell(
-          onTap: onTap,
-          borderRadius: BorderRadius.circular(9),
-          child: Container(
-            width: 32,
-            height: 32,
-            alignment: Alignment.center,
-            decoration: BoxDecoration(
-              color: const Color(0xCCF7E9D2),
-              borderRadius: BorderRadius.circular(9),
-              border: Border.all(color: const Color(0xAA8E6F48), width: 1),
-            ),
-            child: Icon(icon, size: 17, color: const Color(0xFF7A552C)),
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _profileTestamentToggleButton({
-    required String label,
-    required bool selected,
-    required VoidCallback onTap,
-  }) {
-    return Material(
-      color: Colors.transparent,
-      child: InkWell(
-        onTap: onTap,
-        borderRadius: BorderRadius.circular(10),
-        child: Container(
-          width: 54,
-          height: double.infinity,
-          alignment: Alignment.center,
-          decoration: softButtonDecoration(selected: selected),
-          child: Text(
-            label,
-            maxLines: 1,
-            overflow: TextOverflow.ellipsis,
-            style: TextStyle(
-              color: selected
-                  ? const Color(0xFFFDF8EE)
-                  : const Color(0xFF4A331D),
-              fontSize: 13.2,
-              fontWeight: FontWeight.w900,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _profileTopStatCard({required String title, required String value}) {
-    return Container(
-      constraints: const BoxConstraints(minHeight: 58),
-      decoration: floatingPanelDecoration(
-        color: const Color(0xFFF7E9D2),
-        shadowOpacity: 0.08,
-      ),
-      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 8),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Text(
-            title,
-            maxLines: 2,
-            overflow: TextOverflow.ellipsis,
-            textAlign: TextAlign.center,
-            style: const TextStyle(
-              color: Color(0xFF6A4C2E),
-              fontWeight: FontWeight.w800,
-              fontSize: 13.2,
-              height: 1.15,
-            ),
-          ),
-          const SizedBox(height: 2),
-          Text(
-            value,
-            maxLines: 1,
-            overflow: TextOverflow.ellipsis,
-            style: const TextStyle(
-              color: Color(0xFFB06B25),
-              fontWeight: FontWeight.w900,
-              fontSize: 16.8,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _profilePersonProgressRow({
-    required List<Person> rowPeople,
-    required Set<String> completedEventIds,
-  }) {
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 10),
-      decoration: BoxDecoration(
-        color: const Color(0x88F5E8CF),
-        borderRadius: BorderRadius.circular(10),
-        border: Border.all(color: const Color(0xAA8E6F48), width: 1.0),
-      ),
-      child: Row(
-        children: List.generate(rowPeople.length, (index) {
-          final person = rowPeople[index];
-          final progressData = _profileStudyProgressByPersonId[person.id];
-          final progress = progressData?.fraction ?? 0.0;
-          return Expanded(
-            child: Padding(
-              padding: EdgeInsets.only(
-                right: index == rowPeople.length - 1 ? 0 : 6,
-              ),
-              child: Material(
-                color: Colors.transparent,
-                child: InkWell(
-                  onTap: () => _openProfilePersonOverview(
-                    person: person,
-                    completedEventIds: completedEventIds,
-                  ),
-                  borderRadius: BorderRadius.circular(12),
-                  child: Padding(
-                    padding: const EdgeInsets.symmetric(
-                      horizontal: 2,
-                      vertical: 1,
-                    ),
-                    child: LayoutBuilder(
-                      builder: (context, constraints) {
-                        final width = constraints.maxWidth;
-                        final compact = width < 62;
-                        final stacked = width < 108;
-
-                        return Column(
-                          crossAxisAlignment: CrossAxisAlignment.stretch,
-                          children: [
-                            if (stacked)
-                              Column(
-                                children: [
-                                  PersonAvatar(
-                                    person: person,
-                                    size: compact ? 24 : 26,
-                                  ),
-                                  SizedBox(height: compact ? 4 : 5),
-                                  Text(
-                                    person.name,
-                                    maxLines: 1,
-                                    overflow: TextOverflow.ellipsis,
-                                    textAlign: TextAlign.center,
-                                    style: TextStyle(
-                                      color: Color(0xFF4A331D),
-                                      fontWeight: FontWeight.w800,
-                                      fontSize: compact ? 10.2 : 11.8,
-                                    ),
-                                  ),
-                                ],
-                              )
-                            else
-                              Row(
-                                children: [
-                                  PersonAvatar(person: person, size: 28),
-                                  const SizedBox(width: 6),
-                                  Expanded(
-                                    child: Text(
-                                      person.name,
-                                      maxLines: 1,
-                                      overflow: TextOverflow.ellipsis,
-                                      style: const TextStyle(
-                                        color: Color(0xFF4A331D),
-                                        fontWeight: FontWeight.w800,
-                                        fontSize: 12.6,
-                                      ),
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            const SizedBox(height: 7),
-                            ClipRRect(
-                              borderRadius: BorderRadius.circular(999),
-                              child: LinearProgressIndicator(
-                                minHeight: compact ? 7 : 8,
-                                value: progress,
-                                backgroundColor: const Color(0x664E3A26),
-                                valueColor: const AlwaysStoppedAnimation<Color>(
-                                  Color(0xFFC6922D),
-                                ),
-                              ),
-                            ),
-                          ],
-                        );
-                      },
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          );
-        }),
-      ),
-    );
-  }
-
-  Future<void> _openProfilePersonOverview({
-    required Person person,
-    required Set<String> completedEventIds,
-  }) async {
-    final repo = ref.read(storyRepositoryProvider);
-    final progressData = _profileStudyProgressByPersonId[person.id];
-    final completedCount = progressData?.completedCount ?? 0;
-    final totalCount = progressData?.totalCount ?? 0;
-    final progress = progressData?.fraction ?? 0.0;
-    final eventsFuture = repo.fetchEventsForPerson(person.id);
-
-    await showGeneralDialog<void>(
-      context: context,
-      barrierDismissible: true,
-      barrierLabel: 'close',
-      barrierColor: Colors.black54,
-      transitionDuration: const Duration(milliseconds: 280),
-      pageBuilder: (dialogContext, _, __) {
-        return Center(
-          child: ConstrainedBox(
-            constraints: BoxConstraints(
-              maxWidth: 760,
-              maxHeight: MediaQuery.of(dialogContext).size.height * 0.84,
-              minWidth: 320,
-            ),
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 10),
-              child: Material(
-                color: Colors.transparent,
-                child: Container(
-                  clipBehavior: Clip.hardEdge,
-                  decoration: modalSurfaceDecoration(),
-                  child: Stack(
-                    children: [
-                      Padding(
-                        padding: const EdgeInsets.fromLTRB(18, 18, 18, 18),
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.stretch,
-                          children: [
-                            Row(
-                              crossAxisAlignment: CrossAxisAlignment.center,
-                              children: [
-                                PersonAvatar(person: person, size: 58),
-                                const SizedBox(width: 12),
-                                Expanded(
-                                  child: Column(
-                                    crossAxisAlignment:
-                                        CrossAxisAlignment.start,
-                                    children: [
-                                      Row(
-                                        crossAxisAlignment:
-                                            CrossAxisAlignment.center,
-                                        children: [
-                                          Expanded(
-                                            flex: 4,
-                                            child: Text(
-                                              person.name,
-                                              maxLines: 1,
-                                              overflow: TextOverflow.ellipsis,
-                                              style: const TextStyle(
-                                                color: Color(0xFF3A2B15),
-                                                fontSize: 21,
-                                                fontWeight: FontWeight.w900,
-                                              ),
-                                            ),
-                                          ),
-                                          Expanded(
-                                            flex: 5,
-                                            child: Row(
-                                              children: [
-                                                Container(
-                                                  padding:
-                                                      const EdgeInsets.symmetric(
-                                                        horizontal: 10,
-                                                        vertical: 6,
-                                                      ),
-                                                  decoration:
-                                                      headerChipDecoration(),
-                                                  child: Text(
-                                                    '$completedCount / $totalCount',
-                                                    style: const TextStyle(
-                                                      color: Color(0xFF6A4C2E),
-                                                      fontSize: 11.5,
-                                                      fontWeight:
-                                                          FontWeight.w900,
-                                                    ),
-                                                  ),
-                                                ),
-                                                const SizedBox(width: 8),
-                                                Expanded(
-                                                  child: ClipRRect(
-                                                    borderRadius:
-                                                        BorderRadius.circular(
-                                                          999,
-                                                        ),
-                                                    child: LinearProgressIndicator(
-                                                      minHeight: 8,
-                                                      value: progress,
-                                                      backgroundColor:
-                                                          const Color(
-                                                            0x664E3A26,
-                                                          ),
-                                                      valueColor:
-                                                          const AlwaysStoppedAnimation<
-                                                            Color
-                                                          >(Color(0xFFC6922D)),
-                                                    ),
-                                                  ),
-                                                ),
-                                              ],
-                                            ),
-                                          ),
-                                        ],
-                                      ),
-                                      const SizedBox(height: 10),
-                                      Text(
-                                        ((person.description ?? '')
-                                                    .trim()
-                                                    .isNotEmpty
-                                                ? person.description
-                                                : person.tagline) ??
-                                            '아직 등록된 인물 소개가 없습니다.',
-                                        maxLines: 2,
-                                        overflow: TextOverflow.ellipsis,
-                                        style: const TextStyle(
-                                          color: Color(0xFF4D381F),
-                                          fontSize: 13,
-                                          height: 1.48,
-                                          fontWeight: FontWeight.w700,
-                                        ),
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                                const SizedBox(width: 28),
-                              ],
-                            ),
-                            const SizedBox(height: 12),
-                            const Text(
-                              '사건 목록',
-                              style: TextStyle(
-                                color: Color(0xFF4D381F),
-                                fontSize: 14,
-                                fontWeight: FontWeight.w900,
-                              ),
-                            ),
-                            const SizedBox(height: 8),
-                            Expanded(
-                              child: FutureBuilder<List<StoryEvent>>(
-                                future: eventsFuture,
-                                builder: (context, snapshot) {
-                                  if (snapshot.connectionState ==
-                                      ConnectionState.waiting) {
-                                    return const Center(
-                                      child: CircularProgressIndicator(
-                                        strokeWidth: 2.2,
-                                      ),
-                                    );
-                                  }
-                                  if (snapshot.hasError) {
-                                    return Center(
-                                      child: Text(
-                                        '사건 목록을 불러오지 못했습니다.\n${snapshot.error}',
-                                        textAlign: TextAlign.center,
-                                        style: const TextStyle(
-                                          color: Color(0xFFA63F2D),
-                                          fontWeight: FontWeight.w800,
-                                          height: 1.45,
-                                        ),
-                                      ),
-                                    );
-                                  }
-                                  final events =
-                                      snapshot.data ?? const <StoryEvent>[];
-                                  if (events.isEmpty) {
-                                    return const Center(
-                                      child: Text(
-                                        '등록된 사건이 없습니다.',
-                                        style: TextStyle(
-                                          color: Color(0xFF6D5231),
-                                          fontWeight: FontWeight.w700,
-                                        ),
-                                      ),
-                                    );
-                                  }
-                                  return GridView.builder(
-                                    gridDelegate:
-                                        const SliverGridDelegateWithFixedCrossAxisCount(
-                                          crossAxisCount: 3,
-                                          mainAxisSpacing: 8,
-                                          crossAxisSpacing: 8,
-                                          childAspectRatio: 1.48,
-                                        ),
-                                    itemCount: events.length,
-                                    itemBuilder: (context, index) {
-                                      final event = events[index];
-                                      final isCompleted = completedEventIds
-                                          .contains(event.id);
-                                      final placeText = (event.placeName ?? '')
-                                          .trim();
-                                      final yearText =
-                                          event.startYear?.toString() ?? '-';
-                                      final metaText = placeText.isEmpty
-                                          ? yearText
-                                          : '$placeText · $yearText';
-                                      final summary =
-                                          (event.shortStory ??
-                                                  event.story ??
-                                                  event.summary ??
-                                                  '')
-                                              .trim();
-
-                                      return Material(
-                                        color: Colors.transparent,
-                                        child: InkWell(
-                                          onTap: () {
-                                            Navigator.of(dialogContext).pop();
-                                            widget.onOpenEventDetail(event);
-                                          },
-                                          borderRadius: BorderRadius.circular(
-                                            16,
-                                          ),
-                                          child: Container(
-                                            padding: const EdgeInsets.fromLTRB(
-                                              12,
-                                              10,
-                                              12,
-                                              10,
-                                            ),
-                                            decoration: BoxDecoration(
-                                              color: isCompleted
-                                                  ? const Color(0xFFF3E0BE)
-                                                  : const Color(0xEEF7EBD8),
-                                              borderRadius:
-                                                  BorderRadius.circular(16),
-                                              border: Border.all(
-                                                color: isCompleted
-                                                    ? const Color(0xD2C78956)
-                                                    : const Color(0xB58E6F48),
-                                                width: 1.0,
-                                              ),
-                                            ),
-                                            child: Column(
-                                              crossAxisAlignment:
-                                                  CrossAxisAlignment.start,
-                                              children: [
-                                                Row(
-                                                  children: [
-                                                    Container(
-                                                      width: 24,
-                                                      height: 24,
-                                                      alignment:
-                                                          Alignment.center,
-                                                      decoration: BoxDecoration(
-                                                        color: isCompleted
-                                                            ? const Color(
-                                                                0xFFC8863B,
-                                                              )
-                                                            : const Color(
-                                                                0xFFF4ECDE,
-                                                              ),
-                                                        shape: BoxShape.circle,
-                                                        border: Border.all(
-                                                          color: isCompleted
-                                                              ? const Color(
-                                                                  0xFFF1D39C,
-                                                                )
-                                                              : const Color(
-                                                                  0xBC9A7A4C,
-                                                                ),
-                                                          width: 1.0,
-                                                        ),
-                                                      ),
-                                                      child: Icon(
-                                                        isCompleted
-                                                            ? Icons
-                                                                  .check_rounded
-                                                            : Icons
-                                                                  .circle_outlined,
-                                                        size: isCompleted
-                                                            ? 14
-                                                            : 11.5,
-                                                        color: isCompleted
-                                                            ? const Color(
-                                                                0xFFFDF8EE,
-                                                              )
-                                                            : const Color(
-                                                                0xFF8A6A46,
-                                                              ),
-                                                      ),
-                                                    ),
-                                                    const SizedBox(width: 8),
-                                                    Expanded(
-                                                      child: Text(
-                                                        isCompleted
-                                                            ? '완료'
-                                                            : '미완료',
-                                                        maxLines: 1,
-                                                        overflow: TextOverflow
-                                                            .ellipsis,
-                                                        textAlign:
-                                                            TextAlign.right,
-                                                        style: TextStyle(
-                                                          color: isCompleted
-                                                              ? const Color(
-                                                                  0xFFB26D26,
-                                                                )
-                                                              : const Color(
-                                                                  0xFF8A6A46,
-                                                                ),
-                                                          fontSize: 10.5,
-                                                          fontWeight:
-                                                              FontWeight.w900,
-                                                        ),
-                                                      ),
-                                                    ),
-                                                  ],
-                                                ),
-                                                const SizedBox(height: 8),
-                                                Text(
-                                                  event.title,
-                                                  maxLines: 2,
-                                                  overflow:
-                                                      TextOverflow.ellipsis,
-                                                  style: const TextStyle(
-                                                    color: Color(0xFF3D2D18),
-                                                    fontSize: 13,
-                                                    fontWeight: FontWeight.w900,
-                                                    height: 1.2,
-                                                  ),
-                                                ),
-                                                const SizedBox(height: 4),
-                                                Text(
-                                                  metaText,
-                                                  maxLines: 1,
-                                                  overflow:
-                                                      TextOverflow.ellipsis,
-                                                  style: const TextStyle(
-                                                    color: Color(0xFF7A5E38),
-                                                    fontSize: 10.5,
-                                                    fontWeight: FontWeight.w800,
-                                                  ),
-                                                ),
-                                                if (summary.isNotEmpty) ...[
-                                                  const SizedBox(height: 6),
-                                                  Expanded(
-                                                    child: Text(
-                                                      summary,
-                                                      maxLines: 3,
-                                                      overflow:
-                                                          TextOverflow.ellipsis,
-                                                      style: const TextStyle(
-                                                        color: Color(
-                                                          0xFF5A4326,
-                                                        ),
-                                                        fontSize: 10.6,
-                                                        height: 1.35,
-                                                      ),
-                                                    ),
-                                                  ),
-                                                ] else
-                                                  const Spacer(),
-                                              ],
-                                            ),
-                                          ),
-                                        ),
-                                      );
-                                    },
-                                  );
-                                },
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                      Positioned(
-                        right: 12,
-                        top: 12,
-                        child: modalCloseButton(
-                          onTap: () => Navigator.of(dialogContext).pop(),
-                          size: 32,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-          ),
-        );
-      },
     );
   }
 }
